fix(morgan_vm): Fix warnings + Add define header file

+++ corewar/src/init/parse_args.c
@@ -17,7 +17,7 @@ static int handle_parameter(vm_t *vm, char **av, size_t *i)
 
 int parse_args(vm_t *vm, int ac, char **av)
 {
-    for (size_t i = 1; i < ac;) {
+    for (size_t i = 1; (int)i < ac;) {
         if (handle_parameter(vm, av, &i)) {
             write(2, "Invalid parameter.\n", 19);
             return 84;
